Create abstraction for member commands
